+++ source/Classes/Services/ApigeeFunctions.m
@@ -329,7 +329,7 @@ int Apigee_logging_level()
 {
     ApigeeActiveSettings *activeSettings = Apigee_active_settings();
     if (activeSettings) {
-        return activeSettings.logLevelToMonitor;
+        return (int) activeSettings.logLevelToMonitor;
     } else {
         return 0;
     }
@@ -510,7 +510,7 @@ int Apigee_connected_via_mobile()
 
 
 /***********************  custom config parameters  ***************************/
-int Apigee_number_custom_config_parameters()
+NSUInteger Apigee_number_custom_config_parameters()
 {
     ApigeeActiveSettings *activeSettings = Apigee_active_settings();
     if (activeSettings) {
